requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
WTForms==3.1.1
email-validator==2.2.0
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
yt-dlp==2023.12.30
torch>=2.0.0
google-generativeai==0.3.2
git+https://github.com/openai/whisper.git
python-docx==1.1.2 